JCF: to help test Issue #21404 , add ToySimulator support for generated_fragments_per_event
First thing to be aware of: this change is completely backwards-compatible.

If you add a parameter "generated_fragments_per_event: <N>" to the
FHiCL document for a ToySimulator, for each call to getNext_ it will
return <N> fragments, rather than one and only one fragment.

For backwards compatibility, the first fragment in the ToySimulator
always has the fragment_id provided in the FHiCL. The remaining ones
are this value + 10000. This should help avoid fragment id collisions
with other boardreaders, though it's not a guarantee against
them. Also if generated_fragments_per_event > 1, there may be problems
if a routingmaster is used.