JCF: in ToySimulator, when running with one event per subrun have getNext_ return the data fragment with sequence ID <N> and the end-of-subrun fragment with sequence ID <N+1>
This is based on work Eric and I did on the novabeamlinedaq teststand
back on Friday. I'd been observing that the subrun wouldn't always
increment for each new event; by switching from having getNext_ return
a data fragment and end-of-subrun fragment with the same sequence ID
(with the end-of-subrun fragment being skipped on the first call to
getNext_) instead we implemented the same logic found in this commit,
and it solved the problem.